requirements.txt
Flask==2.2.5
flask-cors==4.0.0
mysql-connector-python==8.1.0
pandas==2.1.0
xlsxwriter==3.1.6
openpyxl==3.1.2